Renaming and add var for using connection string
