+++ .secrets.baseline
@@ -4,7 +4,7 @@
     "files": "package-lock.json|^.secrets.baseline$",
     "lines": null
   },
-  "generated_at": "2021-05-14T08:47:56Z",
+  "generated_at": "2021-05-18T12:39:21Z",
   "plugins_used": [
     {
       "name": "AWSKeyDetector"
@@ -63,7 +63,7 @@
       {
         "hashed_secret": "d28a460f49cb7ae6e5e99ab7356b7aeb6faf62b0",
         "is_verified": false,
-        "line_number": 51,
+        "line_number": 52,
         "type": "Secret Keyword"
       }
     ]
