+++ app/config/blob-storage.js
@@ -5,7 +5,8 @@ const schema = Joi.object({
   env: Joi.string().valid('development', 'test', 'production').default('development'),
   connectionStr: Joi.string().when('env', { is: 'production', then: Joi.allow('').optional(), otherwise: Joi.required() }),
   storageAccountName: Joi.string().when('env', { is: 'production', then: Joi.required(), otherwise: Joi.allow('').optional() }),
-  containerName: Joi.string().required()
+  containerName: Joi.string().required(),
+  useConnectionStr: Joi.boolean().default(false)
 })
 
 // Build config
@@ -13,7 +14,8 @@ const config = {
   env: process.env.NODE_ENV,
   connectionStr: process.env.BLOB_STORAGE_CONNECTION_STRING,
   storageAccountName: process.env.BLOB_STORAGE_ACCOUNT_NAME,
-  containerName: process.env.BLOB_STORAGE_CONTAINER_NAME
+  containerName: process.env.BLOB_STORAGE_CONTAINER_NAME,
+  useConnectionStr: process.env.USE_BLOB_STORAGE_CONNECTION_STRING
 }
 
 console.log(config)
