+++ helm/ffc-grants-file-sender/templates/config-map.yaml
@@ -12,4 +12,5 @@ data:
   SHAREPOINT_SITE_ID: {{ .Values.container.sharePointSiteId | quote }}
   SHAREPOINT_TENANT_ID: {{ .Values.container.sharePointTenantId | quote }}
   SHAREPOINT_DOCUMENT_LIBRARY_ID: {{ .Values.container.sharePointDocumentLibraryId | quote }}
+  USE_BLOB_STORAGE_CONNECTION_STRING: {{ .Values.container.useBlobStorageConnectionString | quote }}
 {{- end -}}
